+++ src/main/java/ru/erulaev/restaurantvoting/user/repository/MenuRepository.java
@@ -11,9 +11,9 @@ import java.util.Optional;
 @Transactional(readOnly = true)
 public interface MenuRepository extends FoodBaseRepository<Menu> {
 
-    List<Menu> getAllByRestaurantId(long restaurantId);
+    List<Menu> getAllByRestaurantId(int restaurantId);
 
-    Optional<Menu> getWithDishesByDate(long restaurantId, LocalDate date);
+    Optional<Menu> getWithDishesByDate(int restaurantId, LocalDate date);
 
-    Optional<Menu> getByRestaurantIdAndDate(long restaurantId, LocalDate date);
+    Optional<Menu> getByRestaurantIdAndDate(int restaurantId, LocalDate date);
 }
