+++ src/main/java/ru/erulaev/restaurantvoting/user/service/FoodService.java
@@ -5,15 +5,15 @@ import java.util.List;
 
 public interface FoodService<E, T> {
 
-    List<T> getAll(long parentId);
+    List<T> getAll(int parentId);
 
-    T get(long entityId, long parentId);
+    T get(int entityId, int parentId);
 
-    T save(E entity, long parentId);
+    T save(E entity, int parentId);
 
-    void delete(long entityId, long parentId);
+    void delete(int entityId, int parentId);
 
-    default void update(E entity, long entityId, long parentId) {
+    default void update(E entity, int entityId, int parentId) {
         throw new UnsupportedOperationException("Update is unsupported");
     }
 }
